tests/requirements.txt
# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-xdist==3.5.0

# API Clients
requests==2.31.0
httpx==0.25.2

# Supabase
supabase==2.3.2

# Environment Management
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# Assertions & Mocking
pytest-mock==3.12.0
responses==0.24.1

# Reporting
pytest-html==4.1.1
